+++ source/launcher_version.h
@@ -1,6 +1,6 @@
 #ifndef LAUNCHER_VERSION_H
 #define LAUNCHER_VERSION_H
 
-#define LAUNCHER_VERSION_TEXT "v6.6"
+#define LAUNCHER_VERSION_TEXT "v6.7"
 
 #endif